rename mistyped files. clean up code
