+++ src/lib/items/defaultItemRenderer.js
@@ -1,8 +1,8 @@
 import React from 'react'
+import PropTypes from 'prop-types'
 
 export const defaultItemRenderer = ({
   item,
-  timelineContext,
   itemContext,
   getItemProps,
   getResizeProps
@@ -22,4 +22,13 @@ export const defaultItemRenderer = ({
       {itemContext.useResizeHandle ? <div {...rightResizeProps} /> : ''}
     </div>
   )
+}
+
+// TODO: update this to actual prop types. Too much to change before release
+// future me, forgive me.
+defaultItemRenderer.propTypes = {
+  item: PropTypes.any,
+  itemContext: PropTypes.any,
+  getItemProps: PropTypes.any,
+  getResizeProps: PropTypes.any
 }
